Move hashtocurve module into pasta module.
